GameWindow: Set More Variables to Byte for Indexing  <256 values.

+++ Source/GameBeak-Frontend/Classes/GameWindow.cs
@@ -352,22 +352,22 @@ namespace GameBeak_Frontend.Classes
             if (Core.beakGPU.getWindowEnabled())
             {
                 //Draw Window
-                int winX = Core.beakGPU.getWindowX() - 7;
-                int winY = Core.beakGPU.getWindowY();
+                byte winX = (byte)(Core.beakGPU.getWindowX() - 7);
+                byte winY = (byte)(Core.beakGPU.getWindowY());
 
                 bool yTest = (winY <= 144) && ((winY + 256) >= 0);
                 bool xTest = (winX < 160) && ((winX + 256) > 0);
 
                 if (xTest && yTest)
                 {
-                    int xShift = 0;
-                    int yShift = 0;
-                    int x = 0;
-                    int y = 0;
+                    byte xShift = 0;
+                    byte yShift = 0;
+                    byte x = 0;
+                    byte y = 0;
 
                     if (winX < 0)
                     {
-                        xShift = 0 - winX;
+                        xShift = (byte)(0 - winX);
                     }
                     else
                     {
@@ -376,7 +376,7 @@ namespace GameBeak_Frontend.Classes
 
                     if (winY < 0)
                     {
-                        yShift = 0 - winY;
+                        yShift = (byte)(0 - winY);
                     }
                     else
                     {
@@ -399,8 +399,8 @@ namespace GameBeak_Frontend.Classes
             //Draw Sprites
             if (Core.beakGPU.getSpriteEnabled())
             {
-                int x = 0;
-                int y = 0;
+                byte x = 0;
+                byte y = 0;
                 while (y < 144)
                 {
                     if (spritePixels[((x) + ((y) * 256))].a != 0)
@@ -453,22 +453,22 @@ namespace GameBeak_Frontend.Classes
             if (Core.beakGPU.getWindowEnabled())
             {
                 //Draw Window
-                int winX = Core.beakGPU.getWindowX() - 7;
-                int winY = Core.beakGPU.getWindowY();
+                byte winX = (byte)(Core.beakGPU.getWindowX() - 7);
+                byte winY = Core.beakGPU.getWindowY();
 
                 bool yTest = (winY <= 256) && ((winY + 256) >= 0);
                 bool xTest = (winX < 256) && ((winX + 256) > 0);
 
                 if (xTest && yTest)
                 {
-                    int xShift = 0;
-                    int yShift = 0;
-                    int x = 0;
-                    int y = 0;
+                    byte xShift = 0;
+                    byte yShift = 0;
+                    byte x = 0;
+                    byte y = 0;
 
                     if (winX < 0)
                     {
-                        xShift = 0 - winX;
+                        xShift = (byte)(0 - winX);
                     }
                     else
                     {
@@ -477,7 +477,7 @@ namespace GameBeak_Frontend.Classes
 
                     if (winY < 0)
                     {
-                        yShift = 0 - winY;
+                        yShift = (byte)(0 - winY);
                     }
                     else
                     {
@@ -502,8 +502,8 @@ namespace GameBeak_Frontend.Classes
             //Draw Sprites
             if (Core.beakGPU.getSpriteEnabled())
             {
-                int x = 0;
-                int y = 0;
+                byte x = 0;
+                byte y = 0;
                 while (y < 256)
                 {
                     if (spritePixels[((x) + ((y) * 256))].a != 0)
